+++ src/arctickit/types.py
@@ -1,0 +1,5 @@
+from typing import TypeAlias
+
+import polars as pl
+
+PolarsFrame: TypeAlias = pl.DataFrame | pl.LazyFrame